feat(gui): -big refactor -add current percentage on the cursor -made the size of all components proportional to screensize
todo:
-cursor/currentpercentage should be change when the buttons clicked.
-if the drag action is executed slowly, it will not be accepted
